황승민 : 몬스터 LifeBar 구현

+++ DirectX_HSM_KingdomRushFrontiers/GameEngineContents/ContentsData.h
@@ -438,7 +438,7 @@ public:
 
 public:
 	MonsterEnum MonsterType = MonsterEnum::Null;
-	int Hp = 0;
+	float Hp = 0;
 	float AttackRate = -1;
 	int Damage_min = -1;
 	int Damage_MAX = -1;
